+++ client/src/components/Pages/Admin/Interns/index.js
@@ -14,13 +14,16 @@ import * as T from '../../../Common/Typography';
 import Tabs from '../../../Common/Tabs';
 
 import { API_ADMIN_STATS_URL } from '../../../../constants/apiRoutes';
-import { ADMIN_USER_DETAILS } from '../../../../constants/navRoutes';
+import {
+  ADMIN_USER_DETAILS,
+  ADMIN_INTERNS_URL,
+} from '../../../../constants/navRoutes';
 
 import renderExpandedSection from './renderExpandedSection';
 
 const tabs = ['approved', 'approval requests'];
 
-const AdminInterns = () => {
+const AdminInterns = ({ preview }) => {
   const [internRequests, setInternRequests] = useState([]);
   const [approvedInterns, setApprovedInterns] = useState([]);
   const [loading, setLoading] = useState(false);
@@ -83,6 +86,9 @@ const AdminInterns = () => {
         }
         setInternRequests(requests);
         setApprovedInterns(approved);
+        if (preview) {
+          setSelected(1);
+        }
       } catch (err) {
         let errorMsg = 'Something went wrong';
         if (err.response && err.response.status !== 500) {
@@ -93,7 +99,7 @@ const AdminInterns = () => {
     };
     fetchData();
     setLoading(false);
-  }, []);
+  }, [preview]);
 
   const renderTable = () => {
     switch (selected) {
@@ -112,9 +118,12 @@ const AdminInterns = () => {
           <Table
             data={internRequests}
             columns={requestCols}
-            showSearch
+            showSearch={!preview}
             loading={loading}
             expandedSection={renderExpandedSection}
+            tableHeader={preview && 'intern approval requests'}
+            previewLink={preview && ADMIN_INTERNS_URL}
+            previewLinkText={preview && 'View all intern requests'}
           />
         );
       default:
@@ -124,16 +133,25 @@ const AdminInterns = () => {
 
   return (
     <>
-      <Row mb={6}>
-        <Col w={[4, 10, 10]}>
-          <T.H2 color="blue">Interns</T.H2>
-        </Col>
-      </Row>
-      <Row mb={4}>
-        <Col w={[4, 12, 12]}>
-          <Tabs items={tabs} caps handleClick={handleTab} selected={selected} />
-        </Col>
-      </Row>
+      {!preview && (
+        <>
+          <Row mb={6}>
+            <Col w={[4, 10, 10]}>
+              <T.H2 color="blue">Interns</T.H2>
+            </Col>
+          </Row>
+          <Row mb={4}>
+            <Col w={[4, 12, 12]}>
+              <Tabs
+                items={tabs}
+                caps
+                handleClick={handleTab}
+                selected={selected}
+              />
+            </Col>
+          </Row>
+        </>
+      )}
       <Row mb={4}>
         <Col w={[4, 12, 12]}>{renderTable()}</Col>
       </Row>
